fix(vincent-rest-apis): move schemas
